refactor: viewをPageViewディレクトリに変更_#76
