+++ components/pageView/likeComponent.tsx
