+++ components/pageView/postComponent.tsx
