+++ pages/posts/[postsId]/post.tsx
@@ -4,9 +4,9 @@ import { db, postDB } from "../../../firebase";
 import { AppContext } from "../../../context/PageStates";
 import { usePostAction } from "../../../hooks/usePostAciton";
 import { useLikeAction } from "../../../hooks/useLikeAction";
-import PostComponent from "../../../components/view/postComponent";
-import CommentComponent from "../../../components/view/commentComponent";
-import LikeComponent from "../../../components/view/likeComponent";
+import PostComponent from "../../../components/pageView/postComponent";
+import CommentComponent from "../../../components/pageView/commentComponent";
+import LikeComponent from "../../../components/pageView/likeComponent";
 import Layout from "../../../components/layout";
 import styles from "../../../components/scss/post.module.scss";
 import { IconButton } from "@material-ui/core";
